requirements.txt
paddlepaddle-gpu
paddleocr
opencv-python
psycopg2-binary
mlflow
pyyaml
numpy
python-dotenv